[improvement] Refactor code to improve code quality (#46)
## Before this PR
There were some leaky abstractions and a lot of weird logic during code generation.

## After this PR
I've refactored a lot of the core logic hopefully this makes the code a little bit more readable and maintanable. In terms of impact to code gen, the only changes that occurred are that we only import the modules that we need and merge named imports into a single line. For a consolidated view of the code gen changes you can look at a6fda32